# Faster wigner symbols implementation using `math.comb`