--- a/.mailmap
+++ b/.mailmap
@@ -1362,6 +1362,7 @@ Sandeep Murthy <smurthy@protonmail.ch>
 Sandeep Veethu <sandeep.veethu@gmail.com>
 Sanket Agarwal <sanket@sanketagarwal.com>
 Sanya Khurana <sanya@monica.in>
+Saoliang Jin <18322825326@163.com> 0382 <18322825326@163.com>
 Saptarshi Mandal <sapta.iitkgp@gmail.com>
 Saroj Adhikari <adh.saroj@gmail.com>
 Sartaj Singh <singhsartaj94@gmail.com>