--- a/.mailmap
+++ b/.mailmap
@@ -1362,7 +1362,6 @@ Sandeep Murthy <smurthy@protonmail.ch>
 Sandeep Veethu <sandeep.veethu@gmail.com>
 Sanket Agarwal <sanket@sanketagarwal.com>
 Sanya Khurana <sanya@monica.in>
-Saoliang Jin <18322825326@163.com> 0382 <18322825326@163.com>
 Saptarshi Mandal <sapta.iitkgp@gmail.com>
 Saroj Adhikari <adh.saroj@gmail.com>
 Sartaj Singh <singhsartaj94@gmail.com>
@@ -1391,6 +1390,7 @@ Seth Ebner <murgrehk@gmail.com>
 Seth Poulsen <poulsenseth@yahoo.com>
 Seth Troisi <sethtroisi@google.com>
 Shai 'Deshe' Wyborski <shaide@cs.huji.ac.il>
+Shaoliang Jin <18322825326@163.com> 0382 <18322825326@163.com>
 Shashank Agarwal <shashank.agarwal94@gmail.com>
 Shashank KS <shashankks0987@gmail.com> Shashank KS <shashankks@sahaj.ai>
 Shashank Kumar <shashank.kumar.apc13@iitbhu.ac.in>