--- a/.mailmap
+++ b/.mailmap
@@ -1391,6 +1391,7 @@ Seth Poulsen <poulsenseth@yahoo.com>
 Seth Troisi <sethtroisi@google.com>
 Shai 'Deshe' Wyborski <shaide@cs.huji.ac.il>
 Shaoliang Jin <18322825326@163.com> 0382 <18322825326@163.com>
+Shaoliang Jin <18322825326@163.com> jinshaoliang <18322825326@163.com>
 Shashank Agarwal <shashank.agarwal94@gmail.com>
 Shashank KS <shashankks0987@gmail.com> Shashank KS <shashankks@sahaj.ai>
 Shashank Kumar <shashank.kumar.apc13@iitbhu.ac.in>