… sympy/physics/tests/test_clebsch_gordan.py