<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->
No

#### Brief description of what is fixed or changed
The old implementation of wiger symbols was dependent on the global `_Factlist`. Using binomial coefficients allows smaller integers in the summation, so it can speed up the calculations. The formulas can be found in [https://0382.github.io/CGcoefficient.jl/dev/wigner/](https://0382.github.io/CGcoefficient.jl/dev/wigner/).

Benchmarks of the two implementations are done using the following script (Windows 11, Intel Core i7 11800H)
```python
from sympy.physics.wigner import *
from sympy.core.singleton import S
import timeit

def _check_dj_couple(dj1, dj2, dj3):
    """check if three angular momenta couple"""
    if (dj1 + dj2 + dj3) % 2 != 0:
        return False
    if dj1 > dj2 + dj3:
        return False
    if dj2 > dj1 + dj3:
        return False
    if dj3 > dj1 + dj2:
        return False
    return True


def bench_3j(jmax):
    imax = 2 * jmax + 1
    s = 0
    for dj1 in range(imax):
        j1 = dj1 * S.Half
        for dj2 in range(imax):
            j2 = dj2 * S.Half
            for dj3 in range(imax):
                j3 = dj3 * S.Half
                for dm1 in range(-dj1, dj1 + 1):
                    m1 = dm1 * S.Half
                    for dm2 in range(-dj2, dj2 + 1):
                        m2 = dm2 * S.Half
                        m3 = -m1 - m2
                        s += float(wigner_3j(j1, j2, j3, m1, m2, m3))
    return s

def bench_gaunt(lmax):
    s = 0
    for l1 in range(lmax + 1):
        for l2 in range(lmax + 1):
            for l3 in range(lmax + 1):
                for m1 in range(-l1, l1 + 1):
                    for m2 in range(-l2, l2 + 1):
                        m3 = -m1 - m2
                        s += float(gaunt(l1, l2, l3, m1, m2, m3))
    return s

def bench_6j(jmax):
    imax = 2 * jmax + 1
    s = 0
    for dj1 in range(imax):
        j1 = dj1 * S.Half
        for dj2 in range(imax):
            j2 = dj2 * S.Half
            for dj3 in range(imax):
                j3 = dj3 * S.Half
                for dj4 in range(imax):
                    j4 = dj4 * S.Half
                    for dj5 in range(imax):
                        j5 = dj5 * S.Half
                        for dj6 in range(imax):
                            j6 = dj6 * S.Half
                            if not _check_dj_couple(dj1, dj2, dj3):
                                continue
                            if not _check_dj_couple(dj1, dj5, dj6):
                                continue
                            if not _check_dj_couple(dj4, dj2, dj6):
                                continue
                            if not _check_dj_couple(dj4, dj3, dj5):
                                continue
                            s += float(wigner_6j(j1, j2, j3, j4, j5, j6))
    return s

def bench_9j(jmax):
    imax = 2 * jmax + 1
    s = 0
    for dj1 in range(imax):
        j1 = dj1 * S.Half
        for dj2 in range(imax):
            j2 = dj2 * S.Half
            for dj3 in range(imax):
                j3 = dj3 * S.Half
                for dj4 in range(imax):
                    j4 = dj4 * S.Half
                    for dj5 in range(imax):
                        j5 = dj5 * S.Half
                        for dj6 in range(imax):
                            j6 = dj6 * S.Half
                            for dj7 in range(imax):
                                j7 = dj7 * S.Half
                                for dj8 in range(imax):
                                    j8 = dj8 * S.Half
                                    for dj9 in range(imax):
                                        j9 = dj9 * S.Half
                                        if not _check_dj_couple(dj1, dj2, dj3):
                                            continue
                                        if not _check_dj_couple(dj4, dj5, dj6):
                                            continue
                                        if not _check_dj_couple(dj7, dj8, dj9):
                                            continue
                                        if not _check_dj_couple(dj1, dj4, dj7):
                                            continue
                                        if not _check_dj_couple(dj2, dj5, dj8):
                                            continue
                                        if not _check_dj_couple(dj3, dj6, dj9):
                                            continue
                                        s += float(wigner_9j(j1, j2, j3, j4, j5, j6, j7, j8, j9))
    return s

if __name__ == "__main__":
    timeit_result = timeit.timeit('bench_3j(5)', globals=globals(), number=1)
    print(f"Time taken for bench_3j(5): {timeit_result:.6f} seconds")
    timeit_result = timeit.timeit('bench_gaunt(8)', globals=globals(), number=1)
    print(f"Time taken for bench_gaunt(8): {timeit_result:.6f} seconds")
    timeit_result = timeit.timeit('bench_6j(4)', globals=globals(), number=1)
    print(f"Time taken for bench_6j(4): {timeit_result:.6f} seconds")
    timeit_result = timeit.timeit('bench_9j(2)', globals=globals(), number=1)
    print(f"Time taken for bench_9j(2): {timeit_result:.6f} seconds")
```

The old implementation
```text
Time taken for bench_3j(5): 20.075024 seconds
Time taken for bench_gaunt(8): 4.578857 seconds
Time taken for bench_6j(4): 6.226591 seconds
Time taken for bench_9j(2): 65.655845 seconds
```
The new implementation
```text
Time taken for bench_3j(5): 2.081056 seconds
Time taken for bench_gaunt(8): 2.727193 seconds
Time taken for bench_6j(4): 2.509255 seconds
Time taken for bench_9j(2): 4.877028 seconds
```

#### Other comments

1. The old `wigner_6j` implementation relies on `_big_delta_coeff`, which will raise `ValueError("j values must be integer or half integer and fulfill the triangle relation")`. However, in the old `wigner_3j` implementation, it just returns `S.Zero` if the `j values do not fulfill the triangle relation`. Most other packages will not raise an error in such a condition. So in this improved implementation, the `wigner_6j` will not raise such an error.
2. Why use the `math.comb`? Because `sympy.binomial` is lower than `math.comb` in my test. `math.comb` is introduced in Python 3.8, and sympy only supports Python>=3.9, so it is OK to use `math.comb`.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->

* physics.wigner
  * Improve the speed of Clebsch-Gordan coefficients and Wigner symbols.

<!-- END RELEASE NOTES -->