@oscarbenjamin 2025-04-20
The tests seem pretty exhaustive. Are all code paths covered?

You can measure coverage with `pip install pytest-cov` and then
```
pytest --cov=sympy.physics.wigner sympy/physics/test_wigner.py --cov-report=html
```