@0382 2025-04-22
> Are all code paths covered by tests?
> 
> You can measure coverage with `pip install pytest-cov` and then
> 
> ```
> pytest --cov=sympy.physics.wigner sympy/physics/test_wigner.py --cov-report=html
> ```

Since, the new implementation will not ` raise ValueError("j values must be integer or half integer and fulfill the triangle relation")`. The return `S.Zero` in these paths is not tested. I will add new tests for such paths.