@oscarbenjamin 2025-04-23
If I understand correctly you have basically rewritten every function here so the copyright notice should be amended:
https://github.com/sympy/sympy/blob/980e409ee346fb73b4c6437408a0d94250602864/sympy/physics/wigner.py#L34-L52
It should say something like "the original code was from ... but it has been rewritten ...".