@0382 2025-04-29
I have update some docs about the rewritten functions. And there are some updates:

- Add `prec` parameters for `wigner_3j` and `clebsch_gordan` to maintain consistency with other functions in the file.
- The `prec` cannot `drastically speed up the calculation` in sympy in my test. It can speed up the calculation in Sage (I don't know why), but it cannot speed up the calculation in sympy. So I remove the related comments.

The current code coverage is 100%.