@0382 2025-04-29
I think I need some help for the documentation build system. It failed with
```
The docs build failed with the following errors:
/home/runner/work/sympy/sympy/sympy/physics/wigner.py:docstring of sympy.physics.wigner.wigner_6j:87: ERROR: Unexpected indentation. [docutils]
make[1]: *** [Makefile:74: printwarnings] Error 1
make: *** [Makefile:55: html] Error 2
Error: Process completed with exit code 2.
```
I don't know what's going wrong. How can i repeat the error in my person computer?