@smichr 2025-04-29
When I was working on functions intending to handle ints I found that allowing a non-int (e.g. a SymPy Integer) into the calculations slowed things down a lot. Is it possible to keep the old functionality using Factorial, etc... but make sure that that layer at which the calculation is running is always working with int? Maybe the entry level functions could cast everything to 2*n to make sure that only ints are being used and when the final answer comes back it would be sympyfied?