@0382 2025-04-30
> When I was working on functions intending to handle ints I found that allowing a non-int (e.g. a SymPy Integer) into the calculations slowed things down a lot. Is it possible to keep the old functionality using Factorial, etc... but make sure that that layer at which the calculation is running is always working with int? Maybe the entry level functions could cast everything to 2*n to make sure that only ints are being used and when the final answer comes back it would be sympyfied?
> 
> I guess my real question is whether we know what exactly caused it to be so much slower -- working with non-python int or was it something about the approach that was responsible for the improvement.

This revision also uses built-in python int too. The key is that binomials C(n, k) are much less than factorials n!, so the internal arithmetic is much faster. Besides, the summations in the old version involve additions of Rationals, e.g. 
```python
    for kk in range(int(imin), int(imax) + 1):
        den = _Factlist[int(kk - aa - bb - ee)] * \
            _Factlist[int(kk - cc - dd - ee)] * \
            _Factlist[int(kk - aa - cc - ff)] * \
            _Factlist[int(kk - bb - dd - ff)] * \
            _Factlist[int(aa + bb + cc + dd - kk)] * \
            _Factlist[int(aa + dd + ee + ff - kk)] * \
            _Factlist[int(bb + cc + ee + ff - kk)]
        sumres = sumres + Integer((-1) ** kk * _Factlist[kk + 1]) / den
```
While in the revision, it is pure built-in python int arithmatic
```python
    for ii in range(int(imin), int(imax) + 1):
        ti = comb(ii + 1, j123 + 1) * \
           comb(jpm123, ii - j453) * \
           comb(jpm132, ii - j426) * \
           comb(jpm231, ii - j156)
        sumres = ti - sumres
```
That's why even though the old vision use `_Factlist` to cache factorials, it is still slower.