@0382 2025-04-30
The document error is fixed. I also found that in the document of `wigner_3j`, the last item has a bold font, which is different from the other items. I also fixed it.